+++ kcapp-api.go
@@ -95,6 +95,8 @@ func main() {
 	router.HandleFunc("/venue/{id}", controllers.GetVenue).Methods("GET")
 	router.HandleFunc("/venue/{id}/config", controllers.GetVenueConfiguration).Methods("GET")
 	router.HandleFunc("/venue/{id}/spectate", controllers.SpectateVenue).Methods("GET")
+	router.HandleFunc("/venue/{id}/players", controllers.GetRecentPlayers).Methods("GET")
+	router.HandleFunc("/venue/{id}/matches", controllers.GetActiveVenueMatches).Methods("GET")
 
 	router.HandleFunc("/tournament", controllers.NewTournament).Methods("POST")
 	router.HandleFunc("/tournament", controllers.GetTournaments).Methods("GET")